10. 부드러운 카메라이동 추가, Gravity 기초작업, YSort추가, EditScene 마우스 안맞는 버그 수정.

+++ Engine/Include/Device.cpp
@@ -191,8 +191,8 @@ Vector2 JEONG::Device::GetWindowToDeviceRatio() const
 Vector2 JEONG::Device::GetWinSizeVector2() const
 {
 	Vector2 Temp;
-	Temp.x = m_WinSize.Width;
-	Temp.y = m_WinSize.Height;
+	Temp.x = (float)m_WinSize.Width;
+	Temp.y = (float)m_WinSize.Height;
 
 	return Temp;
 }
